requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
django-widget-tweaks==1.5.0
sqlparse==0.5.3
tzdata==2024.2
